Add files via upload
This update includes the training and test functions.  It also includes a simplified data loader that loads the facade data.  The notebook file can be used on COLAB but some changes are needed to be able to pass parameters to the Python scripts directly.  Also, the COLAB notebook should be modified to pull data from this GitHub repository instead of relying on google drive.